refactor package name and update projector page
