+++ .idea/dictionaries/rahul.xml
@@ -6,6 +6,7 @@
       <w>plainsboro</w>
       <w>rahul</w>
       <w>singh</w>
+      <w>wwphsstechdesk</w>
     </words>
   </dictionary>
 </component>
