+++ app/src/main/java/io/wwphsstechdesk/application/Main/ScrollableActivity.java
@@ -1,0 +1,13 @@
+package io.wwphsstechdesk.application.Main;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class ScrollableActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_scrollable);
+    }
+}
