+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.anikaitsingh.myapplication.MainActivity">
+    tools:context="io.wwphsstechdesk.application.Main.MainActivity">
 
     <TextView
         android:id="@+id/myTitle"
@@ -37,7 +37,6 @@
         app:layout_constraintTop_toBottomOf="@+id/myTitle" />
 
     <Button
-        android:id="@+id/FAQ"
         style="@style/Widget.AppCompat.Button.Colored"
         android:layout_width="0dp"
         android:layout_height="48dp"
@@ -67,7 +66,6 @@
         app:layout_constraintTop_toBottomOf="@+id/ContactUs" />
 
     <ImageView
-        android:id="@+id/imageView"
         android:layout_width="567dp"
         android:layout_height="320dp"
         android:layout_marginStart="16dp"
