+++ app/src/main/res/layout/activity_projector.xml
@@ -21,7 +21,7 @@
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="@string/wifi_help"
+            android:text="@string/projector_help"
             android:textAppearance="@style/TextAppearance.AppCompat.Body1" />
     </ScrollView>
 </android.support.constraint.ConstraintLayout>
