+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="colorPrimary">#4caf50</color>
-    <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#f8c146</color>
     <color name="colorTitle">#000000</color>
 </resources>
